Rename metrics instrumentor to task_instrumentor
